% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DerivadaNPuntos.R
\name{cincoPuntos}
\alias{cincoPuntos}
\title{Halla el valor de la derivada en un punto}
\usage{
cincoPuntos(funcion, punto, dato, tipo)
}
\arguments{
\item{punto}{punto en el que se evaluará la derivada}

\item{dato}{puede representar el valor de la variación x(h) o el error con el que se quiere
calcular el valor de la derivada.}

\item{tipo}{especifíca si el anterior parametro(dato) corresponde a cualquiera de los dos datos
que puede representar: TRUE=el dato representa el error, FALSE=el dato representa h.}

\item{expresion}{expresion a la que se le calculará la derivada}
}
\value{
el valor de la derivada de la expresion en el punto dado.
}
\description{
Halla el valor de la derivada por el método de cinco puntos de una expresion en
un punto dado, con la restricción de un error o un h específico(según sea el caso).
}
\examples{
expresion=expression((x^2)+x)
punto=0
dato=1*10^(-5)
cincoPuntos(expresion,punto,dato,FALSE)
}
