% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richDeriv.R
\name{derivadaRichardson}
\alias{derivadaRichardson}
\title{derivadaRichardson(x,y)}
\usage{
derivadaRichardson(x, y, h = 0.01, n = 3)
}
\arguments{
\item{x}{Lista de valores x de una funcion}

\item{y}{Lista de valores y de una funcion}

\item{h}{Valor a incrementar}

\item{n}{Numero de iteraciones}
}
\value{
Derivada de los valores de x y calculo del error
}
\description{
Calcula la derivada respecto a x de los valores X y Y,
exceptuando el primer y último valor de x, depende de h
}
\examples{
x = seq(0, 1, by = 0.1)
y = seq(0, 1, by = 0.1)
datos = derivada(x, y, h = 0.01, n = 3)
datos$resultados
}
