% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrapSon.R
\name{graficarsimpson}
\alias{graficarsimpson}
\title{graficarsimpson}
\usage{
graficarsimpson(func, n, a, b, titulo = NULL, ejex = NULL,
  ejey = NULL)
}
\arguments{
\item{func}{Funcion a graficar por medio del metodo de Simpson}

\item{n}{Numero de particiones para graficar}

\item{a}{Limite inferior de la integral}

\item{b}{Limite superior de la integral}

\item{titulo}{Opcional. Título de la gráfica}

\item{ejex}{Opcional. Eje horizontal de la grafica}

\item{ejey}{opcional. Eje vertical de la grafica}
}
\value{
Imprime una grafica con las aproximaciones hechas mediante el metodo de Simpson
}
\description{
Genera la grafica de la funcion integrada por el metodo de Simpson
}
