% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrapSon.R
\name{graficartrap}
\alias{graficartrap}
\title{graficartrap}
\usage{
graficartrap(func, n, a, b, titulo = NULL, ejex = NULL, ejey = NULL)
}
\arguments{
\item{func}{Funcion a graficar por medio del metodo de Simpson}

\item{n}{Numero de particiones para graficar}

\item{a}{Limite inferior de la integral}

\item{b}{Limite superior de la integral}

\item{titulo}{Opcional. Título de la gráfica}

\item{ejex}{Opcional. Eje horizontal de la grafica}

\item{ejey}{opcional. Eje vertical de la grafica}
}
\value{
Genera la grafica de los trapecios correspondientes a la integral
}
\description{
Calcula la integral de acuerdo a los lineamientos del metodo
de Simpson (Parabolas)
}
