% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DerivadaNPuntos.R
\name{hallarErrorCincoPuntos}
\alias{hallarErrorCincoPuntos}
\title{Halla el error del método de cinco puntos}
\usage{
hallarErrorCincoPuntos(h, expresion, punto)
}
\arguments{
\item{h}{la magnitud de la variación en x}

\item{expresion}{expresion a la que se le calculará la derivada}

\item{punto}{punto en el que se evaluará la derivada}
}
\value{
error de truncamiento del cálculo de la derivada con el método de cinco puntos
}
\description{
Halla el error de truncamiento al calcular el valor de la derivada
de una expresion en un punto dado teniendo en cuenta la magnitud de la variación en x
que determina la precisión del cálculo (h)
}
\examples{
expresion=expression((x^2)+x)
h=1*10^(-5)
punto=0
hallarErrorCincoPuntos(h,expresion,punto)
}
