% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DerivadaNPuntos.R
\name{hallarHTresPuntos}
\alias{hallarHTresPuntos}
\title{Halla la variación de x del método de tres puntos}
\usage{
hallarHTresPuntos(error, expresion, punto)
}
\arguments{
\item{error}{la magnitud del error que determinará el valor de h}

\item{expresion}{expresion a la que se le calculará la derivada}

\item{punto}{punto en el que se evaluará la derivada}
}
\value{
la variación de x (h) para el que la derivada tiene el error de truncamiento dado
}
\description{
Halla la variación de x necesaria (h), para que se obtenga el error dado
como parametro al calcular el valor de la derivada de una expresion en un punto dado.
}
\examples{
expresion=expression((x^2)+x)
error=1*10^(-5)
punto=0
hallarHTresPuntos(error,expresion,punto)
}
