% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richInteg.R
\name{integralRichardson}
\alias{integralRichardson}
\title{integralRichardson(x,y)}
\usage{
integralRichardson(x, y, maxParticiones = TRUE, particiones = 3)
}
\arguments{
\item{x}{Lista de valores x de una funcion}

\item{y}{Lista de valores y de una funcion}

\item{maxParticiones}{Calcular el número máximo de particiones?}

\item{particiones}{Numero de particiones para calcular si no se quieren hacer las máximas}
}
\value{
integral definida entre x [1] and el valor x [length(x)]
}
\description{
Integral definica calculada en terminos de X y Y
}
\examples{
integral(x, y, maxParticiones = TRUE, Particiones = 3)
}
