% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrapSon.R
\name{lagrange.poly}
\alias{lagrange.poly}
\title{lagrange.poly}
\usage{
lagrange.poly(x, y)
}
\arguments{
\item{x}{Valor en x para el polinomio}

\item{y}{Valor en y para el polinomio}
}
\value{
Función con el tercer polinomio
}
\description{
Calcula el tercer polinomio de lagrange
}
