% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raices.R
\name{posicionFalsa}
\alias{posicionFalsa}
\title{Posicion falsa}
\usage{
posicionFalsa(funcion, x0, x1)
}
\arguments{
\item{funcion}{Ecuacion no lineal definida anteriormente como: nombre_funcion <- function(x)(Ecuacion no lineal en terminos de x)}

\item{x0}{Limite inferior}

\item{x1}{Limite superior}
}
\value{
Una tabla con las aproximaciones de la raiz de la ecuacion no lineal, con los diferentes errores desde 10^-1 a 10^-6
}
\description{
Para una ecuacion no lineal que cumpla la condicion de f(x0)*f(x1)<0 recordando que este metodo es un derivado del metodo de secante.
}
