% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raices.R
\name{puntofijo}
\alias{puntofijo}
\title{Punto fijo}
\usage{
puntofijo(funcion, x0)
}
\arguments{
\item{funcion}{Ecuacion no lineal definida anteriormente como: nombre_funcion <- function(x)(Ecuacion no lineal en terminos de x)}

\item{x0}{Punto enviado por el usuario el cual pertenece a la segunda derivada de la funcion donde posiblemente se encuentre la raiz}
}
\value{
Una tabla con las aproximaciones de la raiz de la ecuacion no lineal, con los diferentes errores desde 10^-1 a 10^-6
}
\description{
Para ecuaciones no lineales donde el punto fijo enviado por el usuario existe en el dominio de la derivada de la funcion, aproximando este valor hasta llegar a una raiz con un error hasta de 10^-6
}
