% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raices.R
\name{raizPuFijPosFSec}
\alias{raizPuFijPosFSec}
\title{Raices con metodo de punto fijo, posicion falsa, secante}
\usage{
raizPuFijPosFSec(funcion, x0 = NULL, x1 = NULL)
}
\arguments{
\item{funcion}{Ecuacion no lineal definida anteriormente como: nombre_funcion <- function(x)(Ecuacion no lineal en terminos de x)}

\item{x0}{Limite menor del intervalo para Secante y PosicionFalsa, para PuntoFijo es el valor inicial del usuario}

\item{x1}{Limite mayor del intervalo para Secante y PosicionFalsa}
}
\value{
Una tabla con las aproximaciones de la raiz de la ecuacion no lineal, con los diferentes errores desde 10^-1 a 10^-6
}
\description{
Desarrollado para ecuaciones no lineales, por medio de una funcion , y ya sea un punto o dos
se generara raices para esa ecuacion, teniendo en cuenta las restricciones de cada funcion descrita anteriormente,
debe escribir la ecuacion en forma de funcion y demas parametros y se calculara evaluando las restricciones
de los metodos para esa ecuacion.
}
