% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RungeKutta3y4.R
\name{rungekutta4}
\alias{rungekutta4}
\title{Runge-Kutta orden 4 (RK4)}
\usage{
rungekutta4(dy, ti, tf, y0, h, graficar = TRUE, numpendientes = 10)
}
\arguments{
\item{dy}{Function: f(x, y) en una ecuación diferencial de la forma y'=f(x, y).}

\item{ti}{Número real: t inicial para la solución.}

\item{tf}{Número real: t final para la solución.}

\item{y0}{Número real: valor inicial, es decir y(ti)=y0.}

\item{h}{Número real: tamaño del paso, no puede ser menor a 10^-4 y debe permitir hacer al menos tres puntos en el intervalo de la solución.}

\item{graficar}{Valor lógico: si es verdadero grafíca, sino no (por defecto es verdadero).}

\item{numpendientes}{Número entero: es la raíz cuadrada del número de pendientes de la gráfica (por defecto es 10).}
}
\value{
t: Vector con los valores en x de la solución.
w: Vector con los valores en y de la solución.
error: Vector con los errores de truncamiento de la solución.
}
\description{
Función que permite solucionar una ecuación diferencial ordinaria de primer orden
usando el método Runge-Kutta de cuarto orden, además gráfica y entrega el error de la solución.
}
\examples{
r<-rungekutta4(function(x, y){x-y}, 0, 2, 1, 0.1)
data.frame (x=r$t, y=r$w, "Error truncamiento"=r$error)
}
\references{
Método desarrollado por C. Runge y M. W. Kutta en 1900.

Análisis Numérico. 10a Ed. Richard L. Burden, J. Douglas Faires y Annette M. Burden. Cengage p. 209.
}
\author{
Jhonny Parra y Laura Donado
}
