% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrapSon.R
\name{simpson}
\alias{simpson}
\title{simpson}
\usage{
simpson(fun, a, b, n)
}
\arguments{
\item{fun}{Funcion a integrar por el metodo de Simpson}

\item{a}{Limite inferior de la integral}

\item{b}{Limite superior de la integral}

\item{n}{Numero de particiones para graficar}
}
\value{
Retorna un valor de tipo entero que representa la suma de
como resultado de la integral
}
\description{
Calcula la integral de acuerdo a los lineamientos del metodo
de Simpson (Parabolas)
}
