% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DerivadaNPuntos.R
\name{tablaDosPuntos}
\alias{tablaDosPuntos}
\title{Tabla de derivadas por dos puntos}
\usage{
tablaDosPuntos(funcion, punto, dato, tipo, cantidadIteraciones)
}
\arguments{
\item{funcion}{Expresion a la que se le calculará la derivada}

\item{punto}{Punto en el que se evaluará la derivada}

\item{dato}{Aqui se contiene h o el error dependiendo del parametro tipo}

\item{tipo}{Este es un valor booleano que sirve para saber a partir de que calcular la derivada:(True es a partir del error)(False a partir de h)}

\item{cantidadIteraciones}{Cuantas veces dividiremos entre dos el valor del dato. Con mas iteraciones, el dato que se encuentra en la ultima tupla es mas exacto, o mas cercano a la respuesta verdadera.}

\item{h}{La magnitud de la variación en x}
}
\value{
No retorna nada, imprime la matriz del error, el h y el valor de la funcion derivada en un punto
}
\description{
Imprime una tabla de la derivada evaluada f'(a) a partir del error o de h
}
