% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DerivadaNPuntos.R
\name{tablaTresPuntos}
\alias{tablaTresPuntos}
\title{Tabla de derivadas por tres puntos}
\usage{
tablaTresPuntos(funcion, punto, dato, tipo, cantidadIteraciones)
}
\arguments{
\item{funcion}{Expresion a la que se le calculará la derivada}

\item{punto}{Punto en el que se evaluará la derivada}

\item{dato}{Aqui se contiene h o el error dependiendo del parametro tipo}

\item{tipo}{Este es un valor booleano que sirve para saber a partir de que calcular la derivada:(True es a partir del error)(False a partir de h)}

\item{cantidadIteraciones}{Cuantas veces dividiremos entre dos el valor del dato. Con mas iteraciones, el dato que se encuentra en la ultima tupla es mas exacto, o mas cercano a la respuesta verdadera.}

\item{h}{La magnitud de la variacion en x}
}
\value{
No retorna nada, imprime la matriz del error, el h y el valor de la funcion derivada en un punto
}
\description{
Imprime una tabla de la derivada evaluada f'(a) a partir del error o de h y la formula de derivada por tres puntos
}
\examples{
expresion=expression(x^5)
punto=2
dato=90
tipo=FALSE
iteraciones=20
tablaTresPuntos(expresion,punto,dato,tipo,iteraciones)
expresion=expression(x^5)
punto=2
dato=90
tipo=TRUE
iteraciones=20
tablaTresPuntos(expresion,punto,dato,tipo,iteraciones)

}
