% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrapSon.R
\name{trapezoid}
\alias{trapezoid}
\title{trapezoid}
\usage{
trapezoid(func, a, b, n)
}
\arguments{
\item{func}{Funcion a graficar por medio del metodo de Simpson}

\item{a}{Limite inferior de la integral}

\item{b}{Limite superior de la integral}

\item{n}{Numero de particiones para graficar}
}
\value{
Retorna un valor de tipo entero que representa la suma de los
trapecios como resultado de la integral
}
\description{
Calcula la integral de acuerdo a los lineamientos del
metodo del trapecio
}
